[Model] Support unquantized BF16 PLE tables in VLLM_PLE_MMAP
Intel AutoRound W4A16 exports pass the Qwen4Exp PLE n-gram table
through from the base checkpoint as plain BF16, with no weight_scale
tensor at all (verified via an HTTP-range read of a real checkpoint's
shard headers: dtype BF16, shape [2500012, 160], 128 shards, 0 scale
entries). VLLM_PLE_MMAP's dtype gate only ever knew F8_E4M3, so these
checkpoints were refused outright at _validate_layer_shards before
any weight ever streamed.

Replaced the bare _FP8_DTYPES dtype -> torch_dtype map with a
_PleDtype descriptor registry (torch_dtype, requires_scale). F8_E4M3
keeps requires_scale=True and every existing scale behavior
untouched: header-read fallback when a family was never streamed, the
streamed-vs-header cross-check, and the fail-closed
weights_streamed/weight_scale_loaded discriminator. BF16 is
registered with requires_scale=False: _attach_table's entire scale
block (discriminator, fallback, cross-check) now runs only for
requires_scale dtypes, gated on the descriptor rather than on
placeholder state, so a BF16 table with weights_streamed=True and
weight_scale_loaded=False — the normal shape of a real load, since
there is no scale to stream — attaches instead of raising.
Qwen4ExpPLELayer._dequantize_embeddings's non-fp8 branch now casts to
output_dtype instead of returning the tensor unchanged, so a BF16
table served under e.g. --dtype float16 doesn't fail late and
unattributably inside a downstream matmul.

New fail-closed case: a no-scale dtype with a weight_scale present on
disk anyway (exporter confusion, e.g. a half fp8-to-bf16 conversion)
is refused in _validate_layer_shards, at both construction time
(validate_shards_for) and attach time.

No new env vars, and no behavior change for existing FP8 checkpoints.
F16 is deliberately left out — no motivating checkpoint exists yet,
and the registry makes it a one-line follow-up. Serving a BF16
checkpoint roughly doubles per-gather cold I/O table-wide versus an
FP8 one of the same shape (2 bytes/column instead of 1) — a
consequence of row width, not something the readahead pre-pass causes
or fixes.

Tests: pytest tests/models/qwen4_exp/ -x -q
--ignore=tests/models/qwen4_exp/test_qsa_amd.py, 226 passed (221
pre-existing + 5 new: BF16 attach/gather/forward value-exact
passthrough, the dequantize-cast unit test, a stray-on-disk-scale
refusal and its well-formed counterpart, and readahead interop at a
320-byte BF16 row width); tests/test_envs.py, 56 passed.

+++ vllm/models/qwen4_exp/nvidia/ple_layer.py
@@ -641,7 +641,9 @@ class Qwen4ExpPLELayer(nn.Module, MambaBase):
         """Dequantize PLE lookup output."""
 
         if not is_fp8(embeddings):
-            return embeddings
+            # Unquantized (e.g. BF16) tables carry no scale to apply — just
+            # cast to the output dtype, same as the fp8 branch's final cast.
+            return embeddings.to(output_dtype)
         weight_scale = self._get_embedding_weight_scale()
         if weight_scale is None:
             raise RuntimeError("FP8 PLE embedding is missing its global scale")
